ACPI/IORT: Check ATS capability in root complex nodes
Root complex node in IORT has a bit telling whether it supports ATS or
not. Store this bit in the IOMMU fwspec when setting up a device, so it
can be accessed later by an IOMMU driver.

Use the negative version at the moment because it's not clear if/how the
bit needs to be integrated in other firmare descriptions. The SMMU has a
feature bit telling if it supports ATS, which might be sufficient in most
systems for deciding whether or not we should enable the ATS capability in
endpoints.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>